refactor(desktop): extract browser ipc
