+++ apps/desktop/src/main/__tests__/main-process-contract-source-helpers.ts
@@ -7,6 +7,7 @@ const REPO_ROOT = resolve(import.meta.dirname, '../../../../..');
 export const MAIN_PROCESS_SOURCE_REPO_PATHS: readonly string[] = [
   'apps/desktop/src/main/main.ts',
   'apps/desktop/src/main/bot-incoming-main.ts',
+  'apps/desktop/src/main/browser-ipc-main.ts',
   'apps/desktop/src/main/context-budget-policy.ts',
   'apps/desktop/src/main/daily-review-main.ts',
   'apps/desktop/src/main/main-window.ts',
